+++ src/generic_functions.cpp
@@ -29,7 +29,7 @@ namespace uuid {
 
 namespace modbus {
 
-Request::Request(uint16_t device, uint8_t function_code, uint32_t timeout_ms,
+Request::Request(uint16_t device, uint8_t function_code, uint16_t timeout_ms,
 		const std::shared_ptr<Response> &response)
 		: device_(device), function_code_(function_code), timeout_ms_(timeout_ms),
 		response_(response) {
